requirements.txt
certifi
chardet
colorzero
distro
gpiozero
idna
ipython
nba-api
olefile
pandas
pillow
pip-chill
python-apt
ratelimiter
rgbmatrix
rpi.gpio
spidev
ssh-import-id
urllib3
